requirements.txt
numpy>=1.26,<2
scipy>=1.11,<2
pyyaml>=6,<7
pandas>=2.2,<3
